+++ CHANGELOG.md
@@ -6,3 +6,8 @@
 
 # v1.0.4
 - Add various Crystal Chronicled improvements and functions to dynamically add waypoints for reminaing effort.
+
+# v1.0.5
+- Add Azj the world turns acheivement infomation/guide/buttons
+- Added additional tab placeholders for some of the other relevant acheivements.
+- Moved Check World Quests button around on Overview Tab.